Add files via upload
My first task as an intern at Code Casa. The task was to create a number-guessing game using Java programming language.

The game includes the following features:
1. I generated a random number between 1 and 100.
2. The user inputs their guess.
3. I provide feedback to the user, indicating whether their guess was a match, higher, or lower.
4. The game continues until the user guesses the correct number.

In addition to these features, I also added some other elements, such as limiting the number of attempts to 5 or 10, and including multiple rounds. The game currently has two rounds - the primary round and a secondary round that the user can choose to continue. At the end of the game, the user's score is displayed and points are awarded based on the number of attempts.

This task was successfully completed using Java.